develop_group2/deployment/frontend/requirements.txt
streamlit
requests
Pillow
tensorflow-cpu